Geordi Deployment
Added option '--process-dependency-links' to pip installation command
to allow package 'geordi' to successfully install.

This option will be deprecated in future versions of pip, so we will
need to remove it if/when we upgrade.

This may also be rendered moot if the owner of geordi updates the
setup file to indicate a newer version of gprof2dot.