Remove gunicorn_workers var
The default for gunicorn_workers here, was 1.  Elsewhere in the STARS
config, gunicorn_workers was set to 9, e.g., in production, or 3, e.g.,
for the development server.  The default in this role, however, was
overwriting the value set elsewhere.  So we ended up with 1 gunicorn
worker in every deployment.  Not good.  So now it's gone!